warnings
git-svn-id: https://dev.csoft.net/agar/trunk@6471 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ compat/asprintf.c
@@ -50,7 +50,7 @@ asprintf(char **ret, const char *fmt, ...)
 	va_start(ap, fmt);
 	size = vsprintf(buf, fmt, ap);
 	va_end(ap);
-	if (size <= buflen) {
+	if ((size_t)size <= buflen) {
 		*ret = buf;
 		return (size);
 	}
